Move EW-only classes into EventWar package.
